+++ termlink/common.py
@@ -18,9 +18,27 @@ from pronto import Ontology
 
 from termlink.models import Coding, Relationship, RelationshipSchema
 
+_SCOPE_TO_INVERSE_SCOPE = {
+    'equivalent_to': 'equivalent_to'
+}
+
 _SCOPE_TO_EQUIVALENCE = {
     'equivalent_to': 'equivalent'
 }
+
+def _to_inverse_scope(scope):
+    """Converts a scope into its inverse scope
+
+    Args:
+        scope: a `pronto.Synonym.scope`
+
+    Returns:
+        an inverse scope
+    """
+    try:
+        return _SCOPE_TO_INVERSE_SCOPE[scope]
+    except KeyError:
+        raise RuntimeError('scope \'%s\' is not supported' % scope)
 
 def _to_equivalence_from_scope(scope):
     """Converts a scope into an equivalence
@@ -105,6 +123,11 @@ def _get_relationships(uri, system):
                 for reference in references:
                     relationship = _to_equivalence_from_scope(scope)
                     yield _to_relationship(term, relationship, ontology[reference], system)
+            if scope in _SCOPE_TO_INVERSE_SCOPE:
+                inverse = _to_inverse_scope(scope)
+                for reference in references:
+                    relationship = _to_equivalence_from_scope(inverse)
+                    yield _to_relationship(ontology[reference], relationship, term, system)
 
 
 def execute(args):